refactor(cli): extract unstable cancellation gate
